Revert "save"
This reverts commit 27330a959577a1457189cc007443a34ceade0da8.

+++ coap/er-coap-13/er-coap-13.c
@@ -1603,7 +1603,7 @@ size_t coap_set_payload(void *packet, const void *payload, size_t length)
   coap_packet_t *const coap_pkt = (coap_packet_t *) packet;
 
   coap_pkt->payload = (uint8_t *) payload;
-  coap_pkt->payload_len = length;
+  coap_pkt->payload_len = (uint16_t)(length);
 
   return coap_pkt->payload_len;
 }
